web/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
anthropic>=0.28.0
PyJWT>=2.8.0
pandas>=2.2.0
openpyxl>=3.1.2
xlsxwriter>=3.2.0
google-auth>=2.29.0
google-api-python-client>=2.131.0
